double -> int

+++ drawit/src/drawit/IntVector.java
@@ -5,7 +5,7 @@ public class IntVector {
     private int x;
     private int y;
 
-    public IntVector(double x, double y) {
+    public IntVector(int x, int y) {
         this.x = x;
         this.y = y;
     }
